backend/requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==4.2.2
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.7
contourpy==1.2.1
cssselect==1.2.0
cycler==0.12.1
dj-rest-auth==6.0.0
Django==4.2
django-allauth==0.63.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fake-useragent==1.5.1
feedparser==6.0.11
Flask==3.0.3
fonttools==4.53.1
fpdf==1.7.2
frozendict==2.4.4
greenlet==3.0.3
gunicorn==20.1.0
html5lib==1.1
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.1
multitasking==0.0.11
mysql-connector-python==9.0.0
mysqlclient==2.1.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
parse==1.20.2
peewee==3.17.6
pillow==10.4.0
platformdirs==4.2.2
prometheus_client==0.20.0
pyee==11.1.0
PyJWT==2.8.0
PyMySQL==1.1.1
pyparsing==3.1.2
pyppeteer==2.0.0
pyquery==2.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.1.0
python-telegram-bot==13.14
pytz==2024.1
reportlab==4.2.2
requests
schedule==1.2.2
sendgrid==6.0.5
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.5.1
tornado==6.1
tqdm==4.66.4
typing==3.6.2
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.19
w3lib==2.2.1
webencodings==0.5.1
websockets==10.4
Werkzeug==3.0.3
yfinance==0.2.41
zipp==3.19.2
